pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataprepenv"
version = "0.1.0"
description = "OpenEnv-compatible data cleaning environment with three graded tasks."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "tabulate>=0.9.0",
    "openenv-core>=0.2.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "gradio>=4.0.0",
]

[tool.setuptools]
py-modules = ["app", "client", "models", "ui"]

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "graders*", "server*", "tasks*"]
